% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad2jy.R
\name{ad2jy}
\alias{ad2jy}
\title{Convert AD to Japanese year}
\usage{
ad2jy(adyear, use_old_name = TRUE)
}
\arguments{
\item{adyear}{vector of AD year}

\item{use_old_name}{TRUE for 2019 = H31, FALSE for 2019 = R1}
}
\value{
japanese year in alphabetical name
}
\description{
Convert AD to Japanese year
}
\examples{
sample_year <- seq(1926, 2020,by = 1)
ad2jy(adyear = sample_year, use_old_name = TRUE)

}
