% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_prefec_eng.R
\name{factor_prefec}
\alias{factor_prefec}
\title{Factorize Japanese prefectures}
\usage{
factor_prefec(prefec)
}
\arguments{
\item{prefec}{vector of Japanese prefectures}
}
\value{
factored vector
}
\description{
Factorize Japanese prefectures
}
