% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfkakko.R
\name{halfkakko}
\alias{halfkakko}
\title{Converts wide parenthesis to halfwitdh}
\usage{
halfkakko(dat)
}
\arguments{
\item{dat}{string contains wide (zenkaku) parenthesis}
}
\value{
string
}
\description{
Converts wide parenthesis to halfwitdh
}
\examples{
\dontrun{
halfkakko("\uff08ABCD\uff09")
}
}
