% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jy2ad.R
\name{jy2ad}
\alias{jy2ad}
\title{Convert Japanese year to AD}
\usage{
jy2ad(jpyear)
}
\arguments{
\item{jpyear}{vector of Japanese years}
}
\value{
numeric vector
}
\description{
Convert Japanese year to AD
}
\examples{
\dontrun{
single_alphabet <- c(paste0("S", seq(59, 63)), paste0("H", seq(1, 11)))

kanji <- c(paste0("昭和", seq(59, 63)), paste0("平成", seq(1, 11)))

single_kanji <- c(paste0("昭 ", seq(59, 63)), paste0("平 ", seq(1, 11)))

zen_single_alphabet <- c(
paste0("Ｓ", c("５９", "６０", "６１", "６２", "６３")),
paste0("Ｈ", seq(1, 11)))

zen_kanji <- c(
paste0("昭和", c("５９", "６０", "６１", "６２", "６３")),
paste0("平成", seq(1, 11)))

zen_single_kanji <- c(
paste0("昭", c("５９", "６０", "６１", "６２", "６３")),
paste0("平", seq(1, 11)))

jy2ad(single_alphabet)
jy2ad(kanji)
jy2ad(single_kanji)
jy2ad(zen_single_alphabet)
jy2ad(zen_kanji)
jy2ad(zen_single_kanji)
}
}
