% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_prefec_eng.R
\name{prefec_jp2en}
\alias{prefec_jp2en}
\title{Convert Japanese prefec names to English}
\usage{
prefec_jp2en(dat)
}
\arguments{
\item{dat}{Prefecture names in \code{tibble} format}
}
\value{
\code{tibble} Anglicized prefecture names
}
\description{
Convert Japanese prefec names to English
}
