% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_prefec_eng.R
\name{trans_prefec_eng}
\alias{trans_prefec_eng}
\title{Convert Japanese prefecture names to English}
\usage{
trans_prefec_eng(dat)
}
\arguments{
\item{dat}{Japanese prefecture names in \code{vector} or \code{tibble}}
}
\value{
Anglicized Japanese prefecture names
}
\description{
Convert Japanese prefecture names to English
}
