% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvj.R
\name{write_csvj}
\alias{write_csvj}
\title{Write csv with Japanese friendly encoding}
\usage{
write_csvj(dat, filename)
}
\arguments{
\item{dat}{data}

\item{filename}{file name ends with .csv}
}
\value{
csv file
}
\description{
Write csv with Japanese friendly encoding
}
\examples{
\dontrun{
write_csvj(mtcars, "mtcars.csv")}
}
